requirements.txt
selenium
requests
python-dotenv
chromedriver-autoinstaller